ETLJira/requirements.txt
aiohttp
orjson
requests
python-dotenv
pydantic